PYIC-7963: Move deduplication and fullname logic into a NameHelper class
