+++ lambdas/build-cri-oauth-request/src/main/java/uk/gov/di/ipv/core/buildcrioauthrequest/BuildCriOauthRequestHandler.java
@@ -334,7 +334,6 @@ public class BuildCriOauthRequestHandler
         var sharedClaims =
                 SharedClaimsHelper.generateSharedClaims(
                         ipvSessionItem.getEmailAddress(), vcs, getAllowedSharedClaimAttrs(cri));
-        sharedClaims.setName(userIdentityService.deduplicateNames(sharedClaims.getName()));
 
         if (cri.equals(F2F)) {
             evidenceRequest =
